#1703 use javax.annotation @Nullable and @Nonnull

+++ utils/src/main/java/org/owasp/dependencycheck/utils/FileUtils.java
@@ -19,6 +19,7 @@ package org.owasp.dependencycheck.utils;
 
 import java.io.Closeable;
 import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -27,6 +28,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.UUID;
 import org.apache.commons.lang3.SystemUtils;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 /**
  * A collection of utilities for processing information about files.
@@ -61,9 +65,10 @@ public final class FileUtils {
      * @param fileName the file name to retrieve the file extension from.
      * @return the file extension.
      */
-    public static String getFileExtension(String fileName) {
-        final String fileExt = FilenameUtils.getExtension(fileName);
-        return null == fileExt || fileExt.isEmpty() ? null : fileExt.toLowerCase();
+    @Nullable
+    public static String getFileExtension(@Nonnull String fileName) {
+        @Nullable final String fileExt = FilenameUtils.getExtension(fileName);
+        return StringUtils.isNoneEmpty(fileExt)? StringUtils.lowerCase(fileExt) : null;
     }
 
     /**
@@ -73,7 +78,13 @@ public final class FileUtils {
      * @param file the File to delete
      * @return true if the file was deleted successfully, otherwise false
      */
-    public static boolean delete(File file) {
+    @Nonnull
+    public static boolean delete(@Nullable File file) {
+        if(file == null) {
+            LOGGER.warn("cannot delete null File");
+            return false;
+        }
+
         final boolean success = org.apache.commons.io.FileUtils.deleteQuietly(file);
         if (!success) {
             LOGGER.debug("Failed to delete file: {}; attempting to delete on exit.", file.getPath());
@@ -90,7 +101,8 @@ public final class FileUtils {
      * @throws IOException thrown when a directory cannot be created within the
      * base directory
      */
-    public static File createTempDirectory(File base) throws IOException {
+    @Nonnull
+    public static File createTempDirectory(@Nullable final File base) throws IOException {
         final File tempDir = new File(base, "dctemp" + UUID.randomUUID().toString());
         if (tempDir.exists()) {
             return createTempDirectory(base);
@@ -108,12 +120,9 @@ public final class FileUtils {
      *
      * @return a String containing the bit bucket
      */
+    @Nonnull
     public static String getBitBucket() {
-        if (SystemUtils.IS_OS_WINDOWS) {
-            return BIT_BUCKET_WIN;
-        } else {
-            return BIT_BUCKET_UNIX;
-        }
+        return SystemUtils.IS_OS_WINDOWS ? BIT_BUCKET_WIN : BIT_BUCKET_UNIX;
     }
 
     /**
@@ -122,7 +131,7 @@ public final class FileUtils {
      *
      * @param closeable to be closed
      */
-    public static void close(Closeable closeable) {
+    public static void close(@Nullable final Closeable closeable) {
         if (null != closeable) {
             try {
                 closeable.close();
@@ -138,7 +147,8 @@ public final class FileUtils {
      * @param resource path
      * @return the input stream for the given resource
      */
-    public static InputStream getResourceAsStream(String resource) {
+    @Nullable
+    public static InputStream getResourceAsStream(@Nonnull String resource) {
         return FileUtils.class.getClassLoader() != null
                 ? FileUtils.class.getClassLoader().getResourceAsStream(resource)
                 : ClassLoader.getSystemResourceAsStream(resource);
